Refactor Kalambury board
