+++ public/locales/en/kalambury.json
@@ -19,5 +19,37 @@
     "play": {
       "guess": "What's on the drawing?"
     }
+  },
+  "board": {
+    "game": {
+      "new_phrase": "Get new phrase"
+    },
+    "guess": {
+      "input": {
+        "placeholder": "The drawing shows...",
+        "action": "Send your guess"
+      }
+    },
+    "wait": {
+      "input": {
+        "placeholder": "Talk to other players while waiting...",
+        "action": "Send your messsage"
+      }
+    }
+  },
+  "sidebar": {
+    "player": {
+      "waiting": "Waiting for player...",
+      "current": "You",
+      "disconnected": "Disconnected",
+      "points": "Points"
+    },
+    "action": {
+      "change": "Changed phrase. Old one was {{phrase}}",
+      "forfeit": "Gave up. The phrase was {{phrase}}",
+      "manage": "Can now manage game.",
+      "draw": "Is now drawing...",
+      "timeout": "Ran out of time. Phrase was {{phrase}}"
+    }
   }
 }
