+++ public/locales/en/lobby.json
@@ -1,0 +1,1 @@
+{}
