+++ src/games/Games.js
@@ -1,4 +1,4 @@
 import { Kalambury } from "./kalambury/Kalambury";
-import KalamburyBoard from "./kalambury/KalamburyBoard";
+import KalamburyBoard from "./kalambury/Board";
 
 export const KalamburyComponent = { game: Kalambury, board: KalamburyBoard };
