+++ src/games/kalambury/sidebar/Action.stories.js
@@ -2,21 +2,21 @@ import React from "react";
 import { action } from "@storybook/addon-actions";
 import { withKnobs, text, boolean } from "@storybook/addon-knobs";
 
-import SidebarAction from "./SidebarAction";
+import Action from "./Action";
 
 export default {
-  component: SidebarAction,
+  component: Action,
   title: "Kalambury/Sidebar/Action",
   excludeStories: /.*Data$/,
 };
 
 export const ActionMessage = () => (
-  <SidebarAction
+  <Action
     action={{ action: "message", text: text("Message text", "Hello there, shall we begin?") }}
   />
 );
 export const ActionGuess = () => (
-  <SidebarAction
+  <Action
     action={{
       action: "guess",
       phrase: text("User guess", "Baba z wozu koniom lżej"),
@@ -26,22 +26,22 @@ export const ActionGuess = () => (
   />
 );
 export const ActionChange = () => (
-  <SidebarAction
+  <Action
     action={{ action: "change", previous: text("Phrase before change", "Baba z wozu koniom lżej") }}
   />
 );
 export const ActionForfeit = () => (
-  <SidebarAction
+  <Action
     action={{
       action: "forfeit",
       previous: text("Phrase before forfeit", "Baba z wozu koniom lżej"),
     }}
   />
 );
-export const ActionManage = () => <SidebarAction action={{ action: "manage" }} />;
-export const ActionDraw = () => <SidebarAction action={{ action: "draw" }} />;
+export const ActionManage = () => <Action action={{ action: "manage" }} />;
+export const ActionDraw = () => <Action action={{ action: "draw" }} />;
 export const ActionTimeout = () => (
-  <SidebarAction
+  <Action
     action={{
       action: "timeout",
       previous: text("Phrase before timeout", "Baba z wozu koniom lżej"),
